Rename directory to date-time-picker
