+++ packages/js/components/src/date-time-picker/index.tsx
@@ -1,0 +1,1 @@
+export * from './date-time-picker';
