--- packages/js/components/src/datetime-picker/index.tsx
@@ -1,1 +1,0 @@
-export * from './datetime-picker';
